Added commit parser
Necessary to parse the XML output from a log command into a records for further use. This builds upon the earlier work where models were added for commits and trailers. There is a lot of process in a commit message which is why there are several steps involved in extracting the right attributes.

The specs for this commit fail only because the `Commands::Log` object hasn't been added yet but will be resolved in the next commit.