Added statistic sanitizer
Necessary to parse text at the end of formatted log output. In other words, the ancillary text which is not supported by log pretty formats. This will allow us to acquire statistical information about the commit.

RuboCop was updated to ignore the missing `else` branch because it'll never be reached if the scan pattern doesn't match because it always answers an empty array which won't be enumerated.

Milestone: minor