Updated log command to include statistics
Necessary for capturing statistical information based on work applied in previous commits where we can use the `--shortstat` option to capture and parse ancillary information which can't be acquired via pretty formatting.

The repository spec was updated to check for zero stats because uncommitted changes aren't computed until _after_ the commit is recorded by Git. The stats are zero across the board in this situation.

Milestone: major