Added directable module
Necessary to extract functionality for use in objects that behave like a commit. This will soon be used by the link:https://alchemists.io/projects/milestoner[Milestoner] gem by using a different data structure but with the same behavior as the `Commit` model in this project. Other projects might benefit from this functionality in the future too.

Milestone: minor