Fixed attributer parser to avoid unnecessary scrubbing
Discovered that indescriminitely scrubbing was causing a massive hit to memory performance so have limited scrubbing to only being performed when there is a legit illegal character. When using link:https://github.com/SamSaffron/memory_profiler[Memory Profiler] here's what I was seeing:

....
140295653 gitt-2.0.1/lib/gitt/parsers/attributer.rb
 23401837 gitt-2.0.1/lib/gitt/parsers/commit.rb
 18408664 gitt-2.0.1/lib/gitt/commands/log.rb
  4151640 gitt-2.0.1/lib/gitt/sanitizers/lines.rb
  3816104 gitt-2.0.1/lib/gitt/models/commit.rb
  3785681 gitt-2.0.1/lib/gitt/parsers/trailer.rb
  3317799 gitt-2.0.1/lib/gitt/sanitizers/statistic.rb
....

The spec subject was updated to be consistent with other specs in this namesapce.

Milestone: patch