Fixed trailer parser to be efficient with empty records
Necessary to reduce then number of unique hashes and trailer models created in memory by only creating trailer records when attributes exist. Otherwise, we'll answer an empty trailer record.

The pattern was updated to document start of line better.

An additional check was added to ensure any content that begins with a comment is immediately ignored. This is necessary because -- starting with Git 2.43.1 -- behavior changed where using `commit-tree` on a `.git/COMMIT_EDITMSG` will now include comments _after_ a trailer and we want to immediately ignore those.

Milestone: patch