Fixed log index parsing with raw content similar to computed statistics
Necessary to improve the robustness of the regular expression when computing statistics while, at the same time, ignoring similar statistics in the raw body. This can happen in rare situations where the commit body (either verbose or not) has content similar to what is matched on via the regular expression when computing statistics. This change reduces the ambiguity.

When the match is a success, the match data is cleaned up further by removing leading and trailing whitespace.

Milestone: patch