Added trailable module
Necessary to enrich objects with additional behavior related to finding and interacting with trailer information. This has been extracted from the `Commit` module in order to apply this same behavior to the `Tag` model in the future. The specs were also extracted as shared examples so this module and models like `Commit` and `Tag` can use the same specs without duplicating effort.

Milestone: minor